app restructure suggestion
