+++ src/components/Filter.js
@@ -1,10 +1,14 @@
 import React from "react";
 
 const capitalize = word => {
-    if(typeof word === 'string') {
-        return word[0].toUpperCase() + word.slice(1)
+  if (typeof word === "string") {
+    if (word) {
+      return word[0].toUpperCase() + word.slice(1);
+    } else {
+      return "Hi";
     }
-}
+  }
+};
 
 const Filter = ({ data, setFilter }) => {
   const arr = data.map(mentor =>
@@ -17,7 +21,11 @@ const Filter = ({ data, setFilter }) => {
     <div className="row col-md-12">
       <h3 className="nav-link">Filters</h3>
       <div className="col-md-12">
-        <select class="form-control dropdown" name="technology" onChange={setFilter}>
+        <select
+          className="form-control dropdown"
+          name="technology"
+          onChange={setFilter}
+        >
           {technologies.map((tec, i) => (
             <option key={i} value={tec} className="dropdown-item">
               {capitalize(tec)}
@@ -26,7 +34,11 @@ const Filter = ({ data, setFilter }) => {
         </select>
       </div>
       <div className="col-md-12">
-        <select class="form-control dropdown" name="country" onChange={setFilter}>
+        <select
+          className="form-control dropdown"
+          name="country"
+          onChange={setFilter}
+        >
           {countries.map((country, i) => (
             <option key={i}>{country}</option>
           ))}
