+++ src/components/Header.js
@@ -29,10 +29,8 @@ class Header extends Component {
       <div className="main thememain-header">
         <div className="container">
           <nav className="navbar navbar-expand-lg navbar-light bg-light thememain-header">
-            <Link to="/">
-              <a className="navbar-brand" href="void" title="Coding-AI">
-                Coding-AI
-              </a>
+            <Link to="/" className="navbar-brand" title="Coding-AI">
+              Coding-AI
             </Link>
             <button
               className="navbar-toggler"
@@ -49,17 +47,15 @@ class Header extends Component {
             <div className="collapse navbar-collapse" id="navbarTogglerDemo02">
               <ul className="navbar-nav mr-auto mt-2 mt-lg-0">
                 <li className="nav-item active">
-                  <Link to="/">
-                    <a className="nav-link" href="void" title="Coding-AI">
-                      Home
-                      <span className="sr-only">(current)</span>
-                    </a>
+                  <Link to="/" className="nav-link" title="Coding-AI">
+                    Home
+                    <span className="sr-only">(current)</span>
                   </Link>
                 </li>
                 <li className="nav-item">
-                  <a className="nav-link" href="void" title="Find a Mentor">
+                  <Link to="/mentors" className="nav-link" title="Find a Mento">
                     Mentors
-                  </a>
+                  </Link>
                 </li>
                 <li className="nav-item">
                   <a className="nav-link disabled" href="void" title="About">
