+++ src/components/MentorItem.js
@@ -1,0 +1,31 @@
+import React, { Component } from "react";
+import { Link } from "react-router-dom";
+import Heart from "./Heart";
+
+class MentorItem extends Component {
+  render() {
+    return (
+      <div className="col-sm-4">
+        <Link
+          to={{ pathname: "/mentor", state: { id: this.props.data } }}
+          className="card"
+        >
+          <div
+            className="thumbnail"
+            style={{ backgroundImage: `url(${this.props.data.image})` }}
+            alt="Card cap"
+          />
+          <div className="content-card">
+            <h3>{this.props.data.name}</h3>
+            <p className="tags">{this.props.data.technology}</p>
+            <div className="bottom-info">
+              <p>{this.props.data.country}</p>
+              <Heart {...this.props.data} />
+            </div>
+          </div>
+        </Link>
+      </div>
+    );
+  }
+}
+export default MentorItem;
