+++ src/components/Router.js
@@ -1,21 +1,25 @@
 import React, { Component } from "react";
-import { Route } from "react-router-dom";
+import { Route, Switch } from "react-router-dom";
 //import logo from "./logo.svg";
 
-import Home from "../components/Home";
-import About from "../components/About";
-import Topics from "../components/Topics";
-import Topic from "../components/Topic";
+import Login from "./Login";
+import Home from "./Home";
+import About from "./About";
+import Topics from "./Topics";
+import Mentors from "./Mentors";
+import Topic from "./Topic";
 
 class Router extends Component {
   render() {
     return (
-      <switch>
+      <Switch>
         <Route path="/" exact component={Home} />
         <Route path="/about" exact component={About} />
+        <Route path="/mentors" exact component={Mentors} />
         <Route path="/topics" exact component={Topics} />
-        <Route path="/topics/:Id" exact component={Topic} />
-      </switch>
+        <Route path="/topics/:id" exact component={Topic} />
+        <Route path="/login" exact component={Login} />
+      </Switch>
     );
   }
 }
