+++ src/index.js
@@ -1,23 +1,23 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter, Route, Switch } from "react-router-dom";
-import Login from "./components/Login";
-import App from './App';
-import * as serviceWorker from './serviceWorker';
+import React from "react";
+import ReactDOM from "react-dom";
+import { BrowserRouter, Switch } from "react-router-dom";
+import Router from "./components/Router";
+import * as serviceWorker from "./serviceWorker";
 
-import 'bootstrap/dist/css/bootstrap.css';
+import "bootstrap/dist/css/bootstrap.css";
+import "./styles.css";
 
 ReactDOM.render(
-        <BrowserRouter>
-            <Switch>
-                <Route path="/login" exact component={Login} />
-                <Route path="/" exact component={App} />
-            </Switch>
-        </BrowserRouter>, 
-        document.getElementById('root')
-    );
+  <BrowserRouter>
+    <Switch>
+      <Router />
+    </Switch>
+  </BrowserRouter>,
+  document.getElementById("root")
+);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: http://bit.ly/CRA-PWA
-serviceWorker.register();
+// serviceWorker.register();
+// Do we need service worker?
